refactor(jangar): extract authority status helpers
